requirements.txt
flask
flask-login
flask-limiter
mistune
PyYAML
mongoengine
voxpopuli
cookie-factory @ git+ssh://git@github.com/loult-elte-fwere/cookie-factory.git#egg=cookie_factory